+++ examples/honeycomb.rs
@@ -11,18 +11,22 @@ use std::time::Duration;
 
 use owiwi::HoneycombConfig;
 use owiwi::Owiwi;
+use owiwi::TraceExporter;
 
 fn main() -> owiwi::Result<()> {
     let api_key = std::env::var("HONEYCOMB_API_KEY").expect("HONEYCOMB_API_KEY must be set");
-
-    let mut owiwi = Owiwi::builder().service_name("honeycomb-example").build();
 
     let config = HoneycombConfig::builder()
         .endpoint("https://api.honeycomb.io".parse().expect("valid URL"))
         .api_key(api_key.into())
         .timeout(Duration::from_secs(5))
         .build();
-    let guard = owiwi.try_init_with(config)?;
+    let mut owiwi = Owiwi::builder()
+        .service_name("honeycomb-example")
+        .trace_exporter(TraceExporter::Honeycomb(config))
+        .build();
+
+    let guard = owiwi.try_init()?;
 
     tracing::info!("sending spans to Honeycomb");